Refactor: Move concurrent utilities to a dedicated package
This commit moves the `ShardingMap` and `ConcurrentCounter` utilities from `pkg/common` to a new, more specific package `pkg/common/concurrent`.

This change improves code organization by grouping related concurrent data structures. All import paths and references have been updated accordingly.
