+++ pkg/common/concurrent/concurrent_counter.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package common
+package concurrent
 
 // A thread-safe counter data structure
 type ConcurrentCounter struct {
